@@ -7,38 +7,40 @@
     "ng": "ng",
     "start": "ng serve",
     "start-open": "ng serve --host 0.0.0.0 --disable-host-check",
-    "build": "ng build",
+    "build": "ng build --aot",
     "test": "ng test",
-    "lint": "ng lint",
+    "lint": "ng lint datacalc-ng --format verbose",
     "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.0.0",
-    "@angular/cdk": "^5.0.0-rc0",
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/forms": "^5.0.0",
-    "@angular/http": "^5.0.0",
-    "@angular/material": "^5.0.0-rc0",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
-    "@angular/platform-server": "^5.0.0",
-    "@angular/router": "^5.0.0",
-    "angularfire2": "^5.0.0-rc.3",
+    "@angular/animations": "^7.2.11",
+    "@angular/cdk": "^7.3.6",
+    "@angular/common": "^7.2.11",
+    "@angular/compiler": "^7.2.11",
+    "@angular/core": "^7.2.11",
+    "@angular/forms": "^7.2.11",
+    "@angular/http": "^7.2.11",
+    "@angular/material": "^7.3.6",
+    "@angular/platform-browser": "^7.2.11",
+    "@angular/platform-browser-dynamic": "^7.2.11",
+    "@angular/platform-server": "^7.2.11",
+    "@angular/router": "^7.2.11",
+    "@angular/fire": "^5.1.2",
     "core-js": "^2.5.0",
-    "firebase": "^4.3.0",
+    "firebase": "^5.5.0",
     "hammerjs": "^2.0.8",
     "immutable": "^3.8.1",
     "lodash": "^4.17.4",
-    "rxjs": "^5.5.2",
+    "rxjs": "^6.4.0",
+    "tslib": "^1.9.0",
     "typed-immutable-record": "0.0.6",
     "zone.js": "^0.8.16"
   },
   "devDependencies": {
-    "@angular/cli": "^1.3.1",
-    "@angular/compiler-cli": "^5.0.0",
+    "@angular-devkit/build-angular": "~0.13.0",
+    "@angular/cli": "^7.3.7",
+    "@angular/compiler-cli": "^7.2.11",
     "@types/jasmine": "2.5.38",
     "@types/lodash": "^4.14.57",
     "@types/node": "~6.0.60",
@@ -54,6 +56,8 @@
     "protractor": "~5.1.0",
     "ts-node": "~2.0.0",
     "tslint": "^5.6.0",
-    "typescript": "2.4.2"
+    "tslint-eslint-rules": "5.4.0",
+    "tslint-sonarts": "^1.6.0",
+    "typescript": "3.2.4"
   }
 }
