@@ -1,24 +1,23 @@
-import {Injectable} from '@angular/core';
-import {ReplaySubject} from 'rxjs/ReplaySubject';
-import {Observable} from 'rxjs/Observable';
-import {AngularFireDatabase} from 'angularfire2/database';
-import {Category} from '../model/category';
-import {Catalog} from '../model/catalog';
-import {Item} from '../model/item';
-import {Map} from 'immutable';
+import {Injectable} from "@angular/core";
+import {ReplaySubject, Observable} from "rxjs";
+import {Category} from "../model/category";
+import {Catalog} from "../model/catalog";
+import {Item} from "../model/item";
+import {Map} from "immutable";
 
-import * as _ from 'lodash';
-import {recordify} from 'typed-immutable-record';
-import {ItemRecord} from '../model/item-record';
-import {CategoryRecord} from '../model/category-record';
+import * as _ from "lodash";
+import {recordify} from "typed-immutable-record";
+import {ItemRecord} from "../model/item-record";
+import {CategoryRecord} from "../model/category-record";
+import {AngularFireDatabase} from "@angular/fire/database";
 
 @Injectable()
 export class CatalogService {
   private catalogObservable: ReplaySubject<Catalog> = new ReplaySubject();
-  private item: Observable<any[]>;
+  private item: Observable<any>;
 
   constructor(ad: AngularFireDatabase) {
-    this.item = ad.object('/catalog').valueChanges();
+    this.item = ad.object("/catalog").valueChanges();
     this.item.subscribe(snapshot => this.buildCatalog(snapshot));
   }
 
