@@ -1,14 +1,14 @@
-import { DatacalcNgPage } from './app.po';
+import { DatacalcNgPage } from "./app.po";
 
-describe('datacalc-ng App', function() {
+describe("datacalc-ng App", () => {
   let page: DatacalcNgPage;
 
   beforeEach(() => {
     page = new DatacalcNgPage();
   });
 
-  it('should display message saying app works', () => {
+  it("should display message saying app works", () => {
     page.navigateTo();
-    expect(page.getParagraphText()).toEqual('app works!');
+    expect(page.getParagraphText()).toEqual("app works!");
   });
 });
