@@ -1,11 +1,11 @@
-import { browser, element, by } from 'protractor';
+import { browser, by, element } from "protractor";
 
 export class DatacalcNgPage {
-  navigateTo() {
-    return browser.get('/');
+  public navigateTo() {
+    return browser.get("/");
   }
 
-  getParagraphText() {
-    return element(by.css('app-root h1')).getText();
+  public getParagraphText() {
+    return element(by.css("app-root h1")).getText();
   }
 }
