@@ -1,17 +1,17 @@
-import {Category} from '../model/category';
-import {Set, Map} from 'immutable';
-import * as _ from 'lodash';
-import {Item} from '../model/item';
-import {Catalog} from '../model/catalog';
+import {Category} from "../model/category";
+import {Set, Map} from "immutable";
+import * as _ from "lodash";
+import {Item} from "../model/item";
+import {Catalog} from "../model/catalog";
 
 export class Util {
   static formatPrice(cents: number): string {
-    return '$' + ( (cents / 100).toFixed(2).replace(/\B(?=(\d{3})+(?!\d))/g, ',') );
+    return "$" + ( (cents / 100).toFixed(2).replace(/\B(?=(\d{3})+(?!\d))/g, ",") );
   }
 
   static getItemsForCategory(catalog: Catalog, categoryId: string): Item[] {
     if (catalog && catalog.entries.has(categoryId)) {
-      return _.sortBy(_.values(catalog.entries.get(categoryId).items), ['id']);
+      return _.sortBy(_.values(catalog.entries.get(categoryId).items), ["id"]);
     }
 
     return [];
@@ -29,11 +29,11 @@ export class Util {
     const orderCategoryItemIds = order ? order.get(categoryId, Set()) : Set();
     const categoryItems = (catalog && catalog.entries.has(categoryId)) ? catalog.entries.get(categoryId).items : {};
 
-    return <Item[]>_.sortBy(
+    return _.sortBy(
       _.values(
         _.pickBy(categoryItems, item => orderCategoryItemIds.contains(item.id))
       ),
-      ['id']);
+      ["id"]);
   }
 
   static getOrderTotal(order: Map<string, Set<string>>, catalog: Catalog): number {
