@@ -1,13 +1,14 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+/* tslint:disable:no-duplicate-string */
+import { async, ComponentFixture, TestBed } from "@angular/core/testing";
 
-import { ItemsComponent } from './items.component';
-import {MatButtonModule, MatCardModule, MatListModule} from '@angular/material';
-import {PricePipe} from '../price.pipe';
-import {By} from '@angular/platform-browser';
-import {Item} from '../model/item';
-import {Component, DebugElement} from '@angular/core';
-import {Set} from 'immutable';
-import {Util} from '../common/Util';
+import { ItemsComponent } from "./items.component";
+import {MatButtonModule, MatCardModule, MatListModule} from "@angular/material";
+import {PricePipe} from "../price.pipe";
+import {By} from "@angular/platform-browser";
+import {Item} from "../model/item";
+import {Component, DebugElement} from "@angular/core";
+import {Set} from "immutable";
+import {Util} from "../common/Util";
 
 @Component({
   template: `
@@ -22,32 +23,32 @@ class TestItemsHostComponent {
 }
 
 function verifyCard(card: DebugElement, expectedItem: Item, cardShouldBeSelected = false) {
-  const cardId = card.nativeElement.attributes['id'].textContent;
+  const cardId = card.nativeElement.attributes.id.textContent;
   expect(cardId).toEqual(expectedItem.id);
-  const cardName = card.query(By.css('.item-title'));
+  const cardName = card.query(By.css(".item-title"));
   expect(cardName.nativeElement.textContent).toEqual(expectedItem.name);
-  const cardPrice = card.query(By.css('.item-price'));
+  const cardPrice = card.query(By.css(".item-price"));
   expect(cardPrice.nativeElement.textContent).toEqual(Util.formatPrice(expectedItem.value));
-  if (expectedItem.hasOwnProperty('description')) {
-    const cardDescription = card.query(By.css('.item-description'));
+  if (expectedItem.hasOwnProperty("description")) {
+    const cardDescription = card.query(By.css(".item-description"));
     expect(cardDescription).toBeTruthy();
     const descriptionText = cardDescription.nativeElement.textContent;
     expectedItem.description.forEach(line => {
       expect(descriptionText).toContain(line);
     });
   }
-  if (expectedItem.hasOwnProperty('note')) {
-    const cardNote = card.query(By.css('.item-note'));
+  if (expectedItem.hasOwnProperty("note")) {
+    const cardNote = card.query(By.css(".item-note"));
     expect(cardNote).toBeTruthy();
     expect(cardNote.nativeElement.textContent).toEqual(expectedItem.note);
   }
-  const cardCommercialSource = card.query(By.css('.item-commercial'));
+  const cardCommercialSource = card.query(By.css(".item-commercial"));
   expect(cardCommercialSource.nativeElement.textContent).toEqual(expectedItem.commercialSource);
-  if (expectedItem.hasOwnProperty('probableSource')) {
-    const cardProbableSource = card.query(By.css('.item-probable'));
+  if (expectedItem.hasOwnProperty("probableSource")) {
+    const cardProbableSource = card.query(By.css(".item-probable"));
     expect(cardProbableSource.nativeElement.textContent).toEqual(expectedItem.probableSource);
   }
-  const cardSelectControl = card.query(By.css('.item-select'));
+  const cardSelectControl = card.query(By.css(".item-select"));
   if (cardShouldBeSelected) {
     expect(cardSelectControl).not.toBeTruthy();
   } else {
@@ -55,32 +56,32 @@ function verifyCard(card: DebugElement, expectedItem: Item, cardShouldBeSelected
   }
 }
 
-describe('ItemsComponent', () => {
+describe("ItemsComponent", () => {
   let component: TestItemsHostComponent;
   let fixture: ComponentFixture<TestItemsHostComponent>;
   const items: Item[] = [
     {
-      id: 'item1',
-      category: 'category',
-      name: 'Item 1',
-      commercialSource: 'Item 1 Source',
+      id: "item1",
+      category: "category",
+      name: "Item 1",
+      commercialSource: "Item 1 Source",
       value: 1234
     },
     {
-      id: 'item2',
-      category: 'category',
-      name: 'Item 2',
-      commercialSource: 'Item 2 Source',
+      id: "item2",
+      category: "category",
+      name: "Item 2",
+      commercialSource: "Item 2 Source",
       value: 4567
     },
     {
-      id: 'item3',
-      category: 'category',
-      name: 'Full Item',
-      commercialSource: 'Item 3 Source',
-      probableSource: 'Item 3 Probable',
-      description: ['Description 1', 'Description 2'],
-      note: 'Item note',
+      id: "item3",
+      category: "category",
+      name: "Full Item",
+      commercialSource: "Item 3 Source",
+      probableSource: "Item 3 Probable",
+      description: ["Description 1", "Description 2"],
+      note: "Item note",
       value: 1234
     }
   ];
@@ -103,59 +104,59 @@ describe('ItemsComponent', () => {
     fixture.detectChanges();
   });
 
-  it('should be created', () => {
+  it("should be created", () => {
     expect(component).toBeTruthy();
   });
 
-  it('should have no items without input', () => {
-    const itemList = fixture.debugElement.queryAll(By.css('.category-item'));
+  it("should have no items without input", () => {
+    const itemList = fixture.debugElement.queryAll(By.css(".category-item"));
     expect(itemList.length).toEqual(0);
   });
 
-  it('should show items with none selected', () => {
+  it("should show items with none selected", () => {
     component.items = items;
     fixture.detectChanges();
-    const itemCards = fixture.debugElement.queryAll(By.css('.category-item'));
+    const itemCards = fixture.debugElement.queryAll(By.css(".category-item"));
     expect(itemCards.length).toEqual(items.length);
     itemCards.forEach((card: DebugElement, index: number) => {
       verifyCard(card, items[index]);
     });
   });
 
-  it('should show items with none selected matching order', () => {
+  it("should show items with none selected matching order", () => {
     component.items = items;
-    component.orderItems = Set(['abc', '123']);
+    component.orderItems = Set(["abc", "123"]);
     fixture.detectChanges();
-    const itemCards = fixture.debugElement.queryAll(By.css('.category-item'));
+    const itemCards = fixture.debugElement.queryAll(By.css(".category-item"));
     expect(itemCards.length).toEqual(items.length);
     itemCards.forEach((card: DebugElement, index: number) => {
       verifyCard(card, items[index]);
     });
   });
 
-  it('should show items with one selected matching order', () => {
+  it("should show items with one selected matching order", () => {
     component.items = items;
-    component.orderItems = Set(['item2']);
+    component.orderItems = Set(["item2"]);
     fixture.detectChanges();
     const expectedSelectIndex = 1;
-    const itemCards = fixture.debugElement.queryAll(By.css('.category-item'));
+    const itemCards = fixture.debugElement.queryAll(By.css(".category-item"));
     expect(itemCards.length).toEqual(items.length);
     itemCards.forEach((card: DebugElement, index: number) => {
       verifyCard(card, items[index], expectedSelectIndex === index);
     });
   });
 
-  it('should send items selected event', () => {
+  it("should send items selected event", () => {
     component.items = items;
     fixture.detectChanges();
-    const itemCards = fixture.debugElement.queryAll(By.css('.category-item'));
+    const itemCards = fixture.debugElement.queryAll(By.css(".category-item"));
     expect(itemCards.length).toEqual(items.length);
     itemCards.forEach((card: DebugElement, index: number) => {
       verifyCard(card, items[index]);
     });
-    const cardSelectControl = itemCards[0].query(By.css('.item-select'));
+    const cardSelectControl = itemCards[0].query(By.css(".item-select"));
     expect(cardSelectControl).toBeTruthy();
-    cardSelectControl.triggerEventHandler('click', 0);
-    expect(component.selectedItem).toEqual('item1');
+    cardSelectControl.triggerEventHandler("click", 0);
+    expect(component.selectedItem).toEqual("item1");
   });
 });
