@@ -1,5 +1,5 @@
-import {Map} from 'immutable';
-import {Category} from './category';
+import {Map} from "immutable";
+import {Category} from "./category";
 
 export interface Catalog {
   readonly entries: Map<string, Category>;
