@@ -1,4 +1,4 @@
-import {TypedRecord} from 'typed-immutable-record';
-import {Category} from './category';
+import {TypedRecord} from "typed-immutable-record";
+import {Category} from "./category";
 
 export interface CategoryRecord extends TypedRecord<CategoryRecord>, Category {}
