@@ -1,4 +1,4 @@
-import {TypedRecord} from 'typed-immutable-record';
-import {Item} from './item';
+import {TypedRecord} from "typed-immutable-record";
+import {Item} from "./item";
 
 export interface ItemRecord extends TypedRecord<ItemRecord>, Item {}
