@@ -1,18 +1,18 @@
 import {
   ChangeDetectionStrategy, Component, ElementRef, EventEmitter, Input, OnChanges, Output,
   SimpleChanges, ViewChild
-} from '@angular/core';
-import {Category} from '../model/category';
+} from "@angular/core";
+import {Category} from "../model/category";
 
-import {Item} from '../model/item';
-import {Set, Map} from 'immutable';
-import {Util} from '../common/Util';
-import {Catalog} from '../model/catalog';
+import {Item} from "../model/item";
+import {Set, Map} from "immutable";
+import {Util} from "../common/Util";
+import {Catalog} from "../model/catalog";
 
 @Component({
-  selector: 'mt-order',
-  templateUrl: './order.component.html',
-  styleUrls: ['./order.component.scss'],
+  selector: "mt-order",
+  templateUrl: "./order.component.html",
+  styleUrls: ["./order.component.scss"],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class OrderComponent implements OnChanges {
@@ -22,7 +22,7 @@ export class OrderComponent implements OnChanges {
 
   @Output() itemRemoved = new EventEmitter<{ categoryId: string, itemId: string }>();
 
-  @ViewChild('orderTotalElement') orderTotalElement: ElementRef;
+  @ViewChild("orderTotalElement") orderTotalElement: ElementRef;
 
   orderTotal = 0;
 
@@ -31,17 +31,17 @@ export class OrderComponent implements OnChanges {
   constructor() {
     // Ugly browser detection hack, but we only want to perform the force redraw for Safari
     const ua = window.navigator.userAgent;
-    this.isSafari = ua.indexOf('Safari') !== -1 && ua.indexOf('Chrome') === -1;
+    this.isSafari = ua.indexOf("Safari") !== -1 && ua.indexOf("Chrome") === -1;
   }
 
   ngOnChanges(changes: SimpleChanges): void {
-    const orderValue = changes.hasOwnProperty('order') ? changes.order.currentValue : this.order;
-    const catalogValue = changes.hasOwnProperty('catalog') ? changes.catalog.currentValue : this.catalog;
+    const orderValue = changes.hasOwnProperty("order") ? changes.order.currentValue : this.order;
+    const catalogValue = changes.hasOwnProperty("catalog") ? changes.catalog.currentValue : this.catalog;
     this.orderTotal = Util.getOrderTotal(orderValue, catalogValue);
-    // Safari won't automatically redraw the total for some reason, so use this ugly hack to force it
+    // Safari won"t automatically redraw the total for some reason, so use this ugly hack to force it
     if (this.isSafari) {
-      this.orderTotalElement.nativeElement.style.display = 'none';
-      setTimeout(() => this.orderTotalElement.nativeElement.style.removeProperty('display'), 0);
+      this.orderTotalElement.nativeElement.style.display = "none";
+      setTimeout(() => this.orderTotalElement.nativeElement.style.removeProperty("display"), 0);
     }
   }
 
