@@ -1,7 +1,7 @@
-import { PricePipe } from './price.pipe';
+import { PricePipe } from "./price.pipe";
 
-describe('PricePipe', () => {
-  it('create an instance', () => {
+describe("PricePipe", () => {
+  it("create an instance", () => {
     const pipe = new PricePipe();
     expect(pipe).toBeTruthy();
   });
