@@ -6,10 +6,10 @@
 export const environment = {
   production: false,
   firebase: {
-    apiKey: 'AIzaSyACjFU8ux-Df1De0gAgBeIDxXMEvafAiQc',
-    authDomain: 'glaring-torch-2436.firebaseapp.com',
-    databaseURL: 'https://glaring-torch-2436.firebaseio.com',
-    storageBucket: 'glaring-torch-2436.appspot.com',
-    messagingSenderId: '844761162138'
+    apiKey: "AIzaSyACjFU8ux-Df1De0gAgBeIDxXMEvafAiQc",
+    authDomain: "glaring-torch-2436.firebaseapp.com",
+    databaseURL: "https://glaring-torch-2436.firebaseio.com",
+    storageBucket: "glaring-torch-2436.appspot.com",
+    messagingSenderId: "844761162138"
   }
 };
